+++ src/types/matin.ts
@@ -1,0 +1,5 @@
+export type NoticeListType = {
+  noticeId: number;
+  content: string;
+  createdAt: string;
+}[];
